requirements.txt
flask
flask-cors
pillow
tensorflow
pytesseract
numpy